+++ packages/react/src/CharacterCount/CharacterCount.test.tsx
@@ -1,11 +1,11 @@
 import { render, screen } from '@testing-library/react'
 import { createRef } from 'react'
-import { FormFieldCharacterCount } from './FormFieldCharacterCount'
+import { CharacterCount } from './CharacterCount'
 import '@testing-library/jest-dom'
 
-describe('Form field character count', () => {
+describe('Character count', () => {
   it('renders', () => {
-    render(<FormFieldCharacterCount length={10} maxLength={100} />)
+    render(<CharacterCount length={10} maxLength={100} />)
 
     const component = screen.getByRole('status')
 
@@ -14,33 +14,33 @@ describe('Form field character count', () => {
   })
 
   it('renders a design system BEM class name', () => {
-    render(<FormFieldCharacterCount length={10} maxLength={100} />)
+    render(<CharacterCount length={10} maxLength={100} />)
 
     const component = screen.getByRole('status')
 
-    expect(component).toHaveClass('ams-form-field-character-count')
+    expect(component).toHaveClass('ams-character-count')
   })
 
   it('renders an additional class name', () => {
-    render(<FormFieldCharacterCount length={10} maxLength={100} className="extra" />)
+    render(<CharacterCount length={10} maxLength={100} className="extra" />)
 
     const component = screen.getByRole('status')
 
-    expect(component).toHaveClass('ams-form-field-character-count extra')
+    expect(component).toHaveClass('ams-character-count extra')
   })
 
   it('renders an error class when length is larger than maxLength', () => {
-    render(<FormFieldCharacterCount length={101} maxLength={100} />)
+    render(<CharacterCount length={101} maxLength={100} />)
 
     const component = screen.getByRole('status')
 
-    expect(component).toHaveClass('ams-form-field-character-count--error')
+    expect(component).toHaveClass('ams-character-count--error')
   })
 
   it('supports ForwardRef in React', () => {
     const ref = createRef<HTMLDivElement>()
 
-    render(<FormFieldCharacterCount length={10} maxLength={100} ref={ref} />)
+    render(<CharacterCount length={10} maxLength={100} ref={ref} />)
 
     const component = screen.getByRole('status')