+++ packages/react/src/FormFieldCharacterCount/index.ts
@@ -1,2 +1,0 @@
-export { FormFieldCharacterCount } from './FormFieldCharacterCount'
-export type { FormFieldCharacterCountProps } from './FormFieldCharacterCount'